Minor improvements in subroutine calculations
Two minor changes have been made in the depths of the code. First,
three parameters in the code have been changed to either be more
consistent with the original paper or to be more consistent with the
original code. Second, nucleation is now calculated by checking if
the cluster radius is less than the upper bound of the current bin,
rather than the bin mean radius.

Signed-off-by: Sebastian D. Eastham <seastham@mit.edu>